PollForSync -> ChangeInfo{}, support pagination.
PollForSync was attached to Route53{}, since I did not want to go
through all the places returning ChangeInfo{} and embed a handle
to Route53. However, there are not that many places, and the new
semantic is much better.

ListRRSets and ListHostedZones will now paginate internally if the
AWS response is truncated. This might be too high level, in which
case we will roll back to a more explicit de-pagination scheme.

This commit should have been 2 commits, but I cannot split a hunk.